# services/backend-java/src/main/java/dev/hafnium/customer/domain/Customer.java
package dev.hafnium.customer.domain;

import jakarta.persistence.*;
import java.time.OffsetDateTime;
import java.util.UUID;
import lombok.AllArgsConstructor;
import lombok.Builder;
import lombok.Data;
import lombok.NoArgsConstructor;
import org.hibernate.annotations.CreationTimestamp;
import org.hibernate.annotations.UpdateTimestamp;

/**
 * Customer entity representing a registered customer in the system.
 *
 * <p>
 * Customers can be either individuals or businesses and go through KYC
 * verification workflows.
 */
@Entity
@Table(name = "customers")
@Data
@Builder
@NoArgsConstructor
@AllArgsConstructor
public class Customer {

    @Id
    @GeneratedValue(strategy = GenerationType.UUID)
    @Column(name = "customer_id")
    private UUID customerId;

    @Column(name = "tenant_id", nullable = false)
    private UUID tenantId;

    @Column(name = "external_id", nullable = false)
    private String externalId;

    @Enumerated(EnumType.STRING)
    @Column(name = "customer_type", nullable = false)
    private CustomerType customerType;

    @Enumerated(EnumType.STRING)
    @Column(name = "status", nullable = false)
    private CustomerStatus status;

    @Enumerated(EnumType.STRING)
    @Column(name = "risk_tier")
    private RiskTier riskTier;

    @Column(name = "metadata", columnDefinition = "jsonb")
    private String metadata;

    @CreationTimestamp
    @Column(name = "created_at", nullable = false, updatable = false)
    private OffsetDateTime createdAt;

    @UpdateTimestamp
    @Column(name = "updated_at", nullable = false)
    private OffsetDateTime updatedAt;

    @Version
    @Column(name = "version")
    private Long version;

    /** Customer type enumeration. */
    public enum CustomerType {
        INDIVIDUAL,
        BUSINESS
    }

    /** Customer status enumeration. */
    public enum CustomerStatus {
        PENDING,
        DOCUMENTS_REQUIRED,
        IN_REVIEW,
        VERIFIED,
        REJECTED
    }

    /** Risk tier classification. */
    public enum RiskTier {
        LOW,
        MEDIUM,
        HIGH
    }
}