# services/backend-java/src/main/java/dev/hafnium/customer/dto/UpdateCustomerRequest.java
package dev.hafnium.customer.dto;

import lombok.AllArgsConstructor;
import lombok.Builder;
import lombok.Data;
import lombok.NoArgsConstructor;

/**
 * DTO for customer update requests.
 */
@Data
@Builder
@NoArgsConstructor
@AllArgsConstructor
public class UpdateCustomerRequest {

    private String status;
    private String riskTier;
    private Object metadata;
}